+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - Update Symfony to latest patch versions !2604 @\_fridtjof_
 - Update deployer to v7.1.3 !2600 @chriswalg
 - Update some backend dependencies !2601 @chriswalg
+- Get rid of includes/setup.php !2609 @\_fridtjof_
 
 
 ## Development documentation